requirements.txt
discord
praw==4.0.0b4
pillow